Studies Project Report
Add a new Studies folder for homework. Add a repository for my personal final project on the Internet Architecture course (TODAI, with Hiroshi Esaki). Project about RL, a brief overview of RL and the specialization of basic algorithm. The repository contains a website that does all the training (algorithm written in JS)